[laravel-vue][counter style dec] bug-fix when edit or create image use url_image or input file , delete image when it not saved on database.

+++ app/Http/Controllers/CounterStyleDeckController.php
@@ -50,19 +50,14 @@ class CounterStyleDeckController extends Controller
         //     'image' =>'required',
         //     'information' => 'required',
         // ]);
-
-        // $request->session()->flash('success', 'Successfull Create Task was successful!');
-        // CounterStyleDecks::create($requestData);
-        // DB::table('counter_style_decks')
-        // ->insert([
-        //     'created_at'=>date('Y-m-d H:i:s'),
-        //     'title'=>$content,
-        // ]);
         $imagePost = '';
-        if($request->file('image')){
+        if($request->url_image !== 'null'){
+            $imagePost = $request->url_image;
+        } else if($request->file('image')){
             $baseUrlImage = 'http://laravel-vue.test/storage/';
             $imagePost = $baseUrlImage . $request->file('image')->store('post-image');
         }
+
         $list_chips_convert = explode(',', $request->list_chips);
         CounterStyleDecks::create([
             'title' => $request->title,
@@ -109,19 +104,30 @@ class CounterStyleDeckController extends Controller
     public function update(Request $request, $id)
     {
         $findData = CounterStyleDecks::where('slug','=',$id)->firstOrFail(); 
-        $imageReplace= str_replace("http://laravel-vue.test/storage/", "", $findData->image);
-        // dd($request);
-        // if($request->file('image')){
-        //     if($imageReplace !== $request->image){
-        //         Storage::delete($imageReplace);
-        //     }
-        // }
+
+        $imagePost = '';
+        $imagePosition = '';
+        if($request->url_image !== 'null'){
+            $imagePost = $request->url_image;
+            $imagePosition = 'new';
+            $this->removeImageOld($request, $findData, $imagePosition);
+        } else if($request->file('image')){
+            $baseUrlImage = 'http://laravel-vue.test/storage/';
+            $imagePost = $baseUrlImage . $request->file('image')->store('post-image');
+            $imagePosition = 'new';
+            $this->removeImageOld($request, $findData, $imagePosition);
+        } else if($request->image !== 'null'){
+            $imagePost= $request->image;
+            $imagePosition = 'old';
+            $this->removeImageOld($request, $findData, $imagePosition);
+        }
+
         $list_chips_convert = explode(',', $request->list_chips);
         CounterStyleDecks::where('id', $findData->id)
             ->update([
             'title' => $request->title,
             'slug' => $request->slug,
-            'image' => $request->image,
+            'image' => $imagePost,
             'information' => $request->information,
             'list_chips' => json_encode($list_chips_convert)
         ]);
@@ -138,14 +144,22 @@ class CounterStyleDeckController extends Controller
     public function destroy($id)
     {
         $findData = CounterStyleDecks::where('slug','=',$id)->firstOrFail();
-        // dd(response()->json($findData->image));
-        $stringManipulate = str_replace('http://laravel-vue.test/storage/','',$findData->image);
-        if($findData->image){
-            $stringManipulate = str_replace('http://laravel-vue.test/storage/','',$findData->image);
-            Storage::delete($stringManipulate);
-        }
+        $findPathImage = stristr($findData->image, 'http://laravel-vue.test/storage/');
+        if($findPathImage){
+            if($findData->image){
+                $stringManipulate = str_replace('http://laravel-vue.test/storage/','',$findData->image);
+                Storage::delete($stringManipulate);
+            }
+        } 
 
         CounterStyleDecks::destroy($findData->id);
         return response()->json(['status'=>true, 'message'=>'Data berhasil dihapus !!!']);
+    }
+
+    public function removeImageOld($request, $findData, $imagePosition){
+        $imageReplace= str_replace("http://laravel-vue.test/storage/", "", $findData->image);
+        if($imagePosition === 'new'){
+            Storage::delete($imageReplace);
+        } 
     }
 }
